Fixing wrongful release of arrow arrays
