+++ src/include/duckdb/common/arrow_wrapper.hpp
@@ -10,7 +10,6 @@
 #include "duckdb/common/arrow.hpp"
 #include "duckdb/common/helper.hpp"
 
-#include "duckdb/main/query_result.hpp"
 //! Here we have the internal duckdb classes that interact with Arrow's Internal Header (i.e., duckdb/commons/arrow.hpp)
 namespace duckdb {
 class ArrowSchemaWrapper {
@@ -39,7 +38,7 @@ public:
 	int64_t number_of_rows;
 	void GetSchema(ArrowSchemaWrapper &schema);
 
-	unique_ptr<ArrowArrayWrapper> GetNextChunk();
+	shared_ptr<ArrowArrayWrapper> GetNextChunk();
 
 	const char *GetError();
 
@@ -48,20 +47,4 @@ public:
 		arrow_array_stream.release = nullptr;
 	}
 };
-
-class ResultArrowArrayStreamWrapper {
-public:
-	explicit ResultArrowArrayStreamWrapper(unique_ptr<QueryResult> result, idx_t approx_batch_size);
-	ArrowArrayStream stream;
-	unique_ptr<QueryResult> result;
-	std::string last_error;
-	idx_t vectors_per_chunk;
-
-private:
-	static int MyStreamGetSchema(struct ArrowArrayStream *stream, struct ArrowSchema *out);
-	static int MyStreamGetNext(struct ArrowArrayStream *stream, struct ArrowArray *out);
-	static void MyStreamRelease(struct ArrowArrayStream *stream);
-	static const char *MyStreamGetLastError(struct ArrowArrayStream *stream);
-};
-
 } // namespace duckdb
