+++ src/include/duckdb/common/types/data_chunk.hpp
@@ -12,6 +12,8 @@
 #include "duckdb/common/types/vector.hpp"
 #include "duckdb/common/winapi.hpp"
 
+#include "duckdb/common/arrow_wrapper.hpp"
+
 struct ArrowArray;
 
 namespace duckdb {
@@ -43,6 +45,8 @@ public:
 
 	//! The vectors owned by the DataChunk.
 	vector<Vector> data;
+	//! If this data chunk is generated through arrow we have to hold the original arrow array
+	shared_ptr<ArrowArrayWrapper> arrow_array;
 
 public:
 	inline idx_t size() const { // NOLINT
