+++ tools/pythonpkg/src/pyresult.cpp
@@ -9,6 +9,7 @@
 #include "duckdb_python/array_wrapper.hpp"
 #include "duckdb/common/types/uuid.hpp"
 #include "duckdb/common/arrow_wrapper.hpp"
+#include "duckdb/common/result_arrow_wrapper.hpp"
 
 namespace duckdb {
